feat(documents): implement full documents workflow
