+++ app/dashboard/documents/new/page.tsx
@@ -3,10 +3,8 @@ import { DocumentNew } from '@/features/documents/document-new'
 
 export default function Page() {
   return (
-    <div className="space-y-6">
-      <Suspense fallback={<div>Loading...</div>}>
-        <DocumentNew />
-      </Suspense>
-    </div>
+    <Suspense fallback={<div>Loading...</div>}>
+      <DocumentNew />
+    </Suspense>
   )
 }
