+++ docs/documents.md
@@ -1,0 +1,12 @@
+# Documents
+
+The documents feature lets you upload, view, edit, and filter project files.
+
+## Upload
+- Navigate to **Dashboard → Documents → Upload Document**.
+- Fill out the form and select a file. Supported types include PDF, DOC, and common image formats.
+
+## Manage
+- Use the filters to search by title, type, or project.
+- View details or download files from each document card.
+- Edit document information from the detail page.
